move schema to be reusable between db and trpc
