+++ packages/project/src/db/style-sources.ts
@@ -5,7 +5,7 @@ import {
   type AppContext,
 } from "@webstudio-is/trpc-interface/server";
 import { StyleSources } from "@webstudio-is/project-build";
-import type { Build, Project } from "./schema";
+import type { Build, Project } from "../shared/schema";
 
 export const patch = async (
   { buildId, projectId }: { buildId: Build["id"]; projectId: Project["id"] },
