+++ packages/project/src/index.ts
@@ -9,7 +9,7 @@ export const utils = {
 } as const;
 
 export * from "./shared/styles";
-export * from "./db/schema";
+export * from "./shared/schema";
 export * from "./shared/canvas-components";
 export type { InstanceInsertionSpec } from "./shared/tree-utils";
 export type { ProjectRouter } from "./trpc";
